requirements.txt
assertpy
boto3
pandas
pymongo
python-dotenv
PyYAML
PyYAML
toolz
